edit all BusView classes to extend React.PureComponent
