+++ source/views/transportation/bus/lib/index.js
@@ -1,2 +1,4 @@
+// @flow
+
 export {getScheduleForNow} from './get-schedule-for-now'
 export {getSetOfStopsForNow} from './get-set-of-stops-for-now'
